+++ lib/open_router/client.rb
@@ -41,7 +41,7 @@ module OpenRouter
       parameters[:stream] = stream if stream
       parameters.merge!(extras)
 
-      json_post(path: "/chat/completions", parameters:).tap do |response|
+      post(path: "/chat/completions", parameters:).tap do |response|
         raise ServerError, "Empty response from OpenRouter. Might be worth retrying once or twice." if response.blank?
         raise ServerError, response.dig("error", "message") if response.dig("error", "message").present?
       end.with_indifferent_access