+++ CHANGELOG.md
@@ -1,5 +1,12 @@
-## [Unreleased]
+## [0.3.0] - 2024-05-03
+
+- Uses Faraday's built-in JSON mode
+- Added support for configuring Faraday and its middleware
+- Spec creates a STDOUT logger by default (headers, bodies, errors)
+- Spec filters Bearer token from logs by default
 
 ## [0.1.0] - 2024-03-19
 
 - Initial release
+
+## [Unreleased]